requirements.txt
PySide6==6.9.1
pyvips==3.0.0
Pillow==11.3.0